restructure gui_management, and its placement
