+++ hints/gui_management/managers/deletion_utils.py
