+++ hints/gui_management/notebook_manager.py
