+++ hints/gui_management/window_manager.py
