+++ hints/tabs/options_tab.py
@@ -6,10 +6,10 @@ from customtkinter import CTkButton, CTkFrame
 from hints.control.program import Program
 from hints.utils.constants import tab_names
 
-from hints.utils.gui_management.creation_utils import CreationUtils
-from hints.utils.gui_management.deletion_utils import DeletionUtils
-from hints.utils.gui_management.reset_utils import ResetUtils
-from hints.utils.gui_management.notebook_manager import NotebookManager
+from hints.gui_management.managers.creation_utils import CreationUtils
+from hints.gui_management.managers.deletion_utils import DeletionUtils
+from hints.gui_management.managers.reset_utils import ResetUtils
+from hints.gui_management.notebook_manager import NotebookManager
 
 
 class OptionsTab:
